46/50 Points Personal Information 5/5 Work Experiences 10/10 Portfolio 10/10 Education 5/5 Skills 5/5 Honors 5/5 References 1/5 Published to gh-pages 5/5
For the References, although the names listed are real, all of the rest of the information is obviously fake.  This is because I don't feel comfortable listing these people as references without asking them first and getting all their details for a school assignment.  Thus, the -4 points.
I also pulled the leader repo before the page was gh-published on Monday, so a lot of the jade files are similar.  I tried to change as much as I could to differentiate them on top of inputting my information, but this structure makes me very uncomfortable due to how easy it is to accidentally grab someone elses code.